% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_processing.R
\name{add_event_info}
\alias{add_event_info}
\title{Add Season, Location, Time, and Event Index to Play-by-Play}
\usage{
add_event_info(data)
}
\arguments{
\item{data}{A data frame containing game data with the following columns:}
}
\value{
A data frame with additional columns for season year, team location, and event index.
}
\description{
This function processes a given dataset by adding season year, team location,
time, and an event index.
}
