% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_game_details}
\alias{add_game_details}
\title{Process Game Details}
\usage{
add_game_details(data)
}
\arguments{
\item{data}{A data frame containing raw data.}
}
\value{
A data frame with processed location data.
}
\description{
This function processes location, game date, and season
}
