% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp.R
\name{add_pbp_points}
\alias{add_pbp_points}
\title{Add Points to Play-by-Play Data}
\usage{
add_pbp_points(data)
}
\arguments{
\item{data}{A data frame with raw play-by-play data}
}
\value{
A data frame with points added to lineups for all play-by-play data
}
\description{
This function processes play-by-play data to add shot points and cumulative
points to lineup data.
}
\details{
' The resulting data frame includes four additional columns:
\describe{
\item{\code{shot_pts_away}}{Tracks the away team points in a single
possession.}
\item{\code{shot_pts_home}}{Tracks the home team points in a single
possession.}
\item{\code{pts_away}}{Tracks the away team cumulative points.}
\item{\code{pts_home}}{Tracks the home team cumulative points.}
}
}
