% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp.R
\name{add_pbp_possessions}
\alias{add_pbp_possessions}
\title{Add Possessions to Play-by-Play Data}
\usage{
add_pbp_possessions(data)
}
\arguments{
\item{data}{A data frame with raw play-by-play data}
}
\value{
A data frame with possessions added to lineups for all play-by-play
}
\description{
This function processes play-by-play data to add possessions to lineup data.
}
\details{
The resulting data frame includes four additional columns:
\describe{
\item{\code{poss}}{Tracks the end of the possession.}
\item{\code{start_poss}}{Tracks the time remaining in the quarter when the
possession started.}
\item{\code{poss_team}}{Tracks the possessing team.}
\item{\code{heave}}{Tracks end of quarter heaves.}
}
}
