% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_schedule_details}
\alias{add_schedule_details}
\title{Add Schedule Details}
\usage{
add_schedule_details(data)
}
\arguments{
\item{data}{A data frame containing NBA stats for teams (must include columns
such as \code{season_year}, \code{team_id}, \code{game_date}, etc.).}
}
\value{
A data frame containing the details for the corresponding season.
}
\description{
This function calculates schedule details (such as back-to-back games,
rest days, etc.) for NBA data.
}
