% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_processing.R
\name{add_team_names}
\alias{add_team_names}
\title{Add Team Names to Play-by-Play}
\usage{
add_team_names(data)
}
\arguments{
\item{data}{A data frame containing game event data.}
}
\value{
A data frame with the original data and two new columns:
\code{home_team_name} and \code{away_team_name}.
}
\description{
This function adds home and away team names to the provided data.
}
