% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_days_next_game}
\alias{calc_days_next_game}
\title{Calculate Days Next Game}
\usage{
calc_days_next_game(game_count, game_date)
}
\arguments{
\item{game_count}{The count of games for the season.}

\item{game_date}{The date of the current game.}
}
\value{
A numeric value representing the days until the next game.
}
\description{
This function calculates the number of days until the next game for a team.
}
