% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_days_rest}
\alias{calc_days_rest}
\title{Calculate Days Rest}
\usage{
calc_days_rest(game_count, game_date)
}
\arguments{
\item{game_count}{The count of games for the season.}

\item{game_date}{The date of the current game.}
}
\value{
A numeric value representing the days of rest.
}
\description{
This function calculates the number of days of rest between games.
}
