% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_game_count}
\alias{calc_game_count}
\title{Calculate Game Count for a Season}
\usage{
calc_game_count()
}
\value{
A numeric vector with game counts.
}
\description{
This function generates a count of games for each team in a season.
}
