% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_is_b2b}
\alias{calc_is_b2b}
\title{Determine If Back-to-Back}
\usage{
calc_is_b2b(days_next_game, days_rest)
}
\arguments{
\item{days_next_game}{The number of days until the next game.}

\item{days_rest}{The number of days of rest before the current game.}
}
\value{
A logical value indicating if it's a back-to-back.
}
\description{
This function determines if the current game is part of a back-to-back (B2B)
scenario.
}
