% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_is_b2b_first}
\alias{calc_is_b2b_first}
\title{Determine If B2B First}
\usage{
calc_is_b2b_first(days_next_game)
}
\arguments{
\item{days_next_game}{The number of days until the next game.}
}
\value{
A logical value indicating if it's the first game of a back-to-back.
}
\description{
This function determines if the current game is the first game of a
back-to-back (B2B) scenario.
}
