% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_is_b2b_second}
\alias{calc_is_b2b_second}
\title{Determine If B2B Second}
\usage{
calc_is_b2b_second(days_rest)
}
\arguments{
\item{days_rest}{The number of days of rest before the current game.}
}
\value{
A logical value indicating if it's the second game of a back-to-back.
}
\description{
This function determines if the current game is the second game of a back-to-back (B2B) scenario.
}
