% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp.R
\name{calculate_advanced_stats}
\alias{calculate_advanced_stats}
\title{Calculate Lineup Advanced Stats}
\usage{
calculate_advanced_stats(lineup_stats)
}
\arguments{
\item{lineup_stats}{A data frame containing play-by-play data with lineups
and stats.}
}
\value{
A tibble containing advanced stats data.
}
\description{
This function calculates the advanced stats for each lineup to be used in
calculating play-by-play box scores.
}
