% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_processing.R
\name{calculate_lineup_playtime}
\alias{calculate_lineup_playtime}
\title{Calculate Lineup Playtime}
\usage{
calculate_lineup_playtime(game_df)
}
\arguments{
\item{game_df}{A data frame containing play-by-play data with lineups.}
}
\value{
A tibble containing playtime data.
}
\description{
This function calculates the playtime for each lineup to be used in
play-by-play box scores.
}
