% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_processing.R
\name{calculate_matchup_stats}
\alias{calculate_matchup_stats}
\title{Calculate Matchup Base Stats}
\usage{
calculate_matchup_stats(game_df, lineup_playtime)
}
\arguments{
\item{game_df}{A data frame of play-by-play events that includes tagged
lineup data.}

\item{lineup_playtime}{A data frame of lineup playtime.}
}
\value{
A tibble containing summed stat columns grouped by game, period,
team, lineup, and opposing lineup.
}
\description{
Calculates base stat totals (e.g., FGM, FGA, AST, REB, etc.) for each
lineup vs. opponent lineup matchup, to be used in play-by-play box score
calculations.
}
