% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_processing.R
\name{calculate_possessions_played}
\alias{calculate_possessions_played}
\title{Calculate Possessions Played per Player}
\usage{
calculate_possessions_played(game_df)
}
\arguments{
\item{game_df}{A data frame with \code{lineup_home}, \code{lineup_away},
and \code{possession}.}
}
\value{
A tibble with player, their role (offense/defense), and
possessions count.
}
\description{
Computes total offensive/defensive possessions for each player.
}
