% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_stats_cols}
\alias{clean_stats_cols}
\title{Clean NBA Stats Data}
\usage{
clean_stats_cols(data)
}
\arguments{
\item{data}{A data frame containing raw NBA stats data.}
}
\value{
A cleaned data frame.
}
\description{
This function cleans the fetched NBA team and player stats data by
removing excess columns.
}
