% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp.R
\name{combine_and_add_players}
\alias{combine_and_add_players}
\title{Combine and Add Missing Players}
\usage{
combine_and_add_players(subs_made, other_players)
}
\arguments{
\item{subs_made}{A data frame containing substitution event data}

\item{other_players}{A data frame containing players involved in events
but not substitutions.}
}
\value{
A data frame with the lineups to start each quarter
}
\description{
This function combines player data from substitutions and other events,
then adds missing player data to the resulting data frame.
}
