% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{combine_team_and_opp_scores}
\alias{combine_team_and_opp_scores}
\title{Combine Team and Opponent Score}
\usage{
combine_team_and_opp_scores(team_all_stats, opp_all_stats)
}
\arguments{
\item{team_all_stats}{A data frame containing processed team stats.}

\item{opp_all_stats}{A data frame containing processed opponent team stats.}
}
\value{
A data frame with combined team and opponent stats.
}
\description{
This function combines team and opponent stats.
}
