% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{convert_height_to_inches}
\alias{convert_height_to_inches}
\title{Convert Height from Feet-Inches to Inches}
\usage{
convert_height_to_inches(height)
}
\arguments{
\item{height}{A character string representing height in "feet-inches" format.}
}
\value{
A numeric value representing height in total inches.
}
\description{
Converts a height given in the format "feet-inches" (e.g., "6-2") to total
inches.
}
