% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{convert_inches_to_meters}
\alias{convert_inches_to_meters}
\title{Convert Height from Inches to Meters}
\usage{
convert_inches_to_meters(height_inches)
}
\arguments{
\item{height_inches}{A numeric value representing height in inches.}
}
\value{
A numeric value representing height in meters.
}
\description{
Converts a height in inches to meters.
}
