% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_minutes_seconds}
\alias{convert_to_minutes_seconds}
\title{Convert Minutes to Minutes:Seconds}
\usage{
convert_to_minutes_seconds(time_numeric)
}
\arguments{
\item{time_numeric}{A numeric value.}
}
\value{
A string value in "MM:SS" format.
}
\description{
Converts a time numeric into minutes and seconds (in MM:SS format).
}
