% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_seconds}
\alias{convert_to_seconds}
\title{Convert Time String to Seconds}
\usage{
convert_to_seconds(time_str)
}
\arguments{
\item{time_str}{A string in "MM:SS" format.}
}
\value{
A numeric value representing the time in seconds.
}
\description{
Converts a time string (in MM:SS format) into seconds.
}
