% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{convert_weight_to_kg}
\alias{convert_weight_to_kg}
\title{Convert Weight from Pounds to Kilograms}
\usage{
convert_weight_to_kg(weight_lbs)
}
\arguments{
\item{weight_lbs}{A numeric value representing weight in pounds.}
}
\value{
A numeric value representing weight in kilograms.
}
\description{
Converts a weight in pounds (lbs) to kilograms (kg).
}
