% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_fanduel.R
\name{fetch_fanduel_data}
\alias{fetch_fanduel_data}
\title{Fetch FanDuel Data from API}
\usage{
fetch_fanduel_data(game_id)
}
\arguments{
\item{game_id}{The ID of the game for which to fetch data.}
}
\value{
A list containing the raw FanDuel data and column names.
}
\description{
This function fetches FanDuel data for a given game ID.
}
