% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_schedule.R
\name{fetch_nba_schedule}
\alias{fetch_nba_schedule}
\title{Fetch NBA Schedule}
\usage{
fetch_nba_schedule()
}
\value{
A list containing the NBA schedule data.
}
\description{
This function fetches NBA schedule data from the NBA API.
}
