% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{fetch_nba_scores}
\alias{fetch_nba_scores}
\title{Fetch NBA Scores from API}
\usage{
fetch_nba_scores(season, season_type)
}
\arguments{
\item{season}{A numeric value representing the season (e.g., 2024).}

\item{season_type}{A character string specifying the type of season
(e.g., "Regular Season").}
}
\value{
A data frame containing raw NBA scores data.
}
\description{
This function fetches NBA scores data for a specified season.
}
