% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_play_by_play.R
\name{fetch_play_by_play_data}
\alias{fetch_play_by_play_data}
\title{Fetch Play-by-Play Data from API}
\usage{
fetch_play_by_play_data(game_id)
}
\arguments{
\item{game_id}{The ID of the game for which to fetch data.}
}
\value{
A list containing the raw play-by-play data and column names.
}
\description{
This function fetches play-by-play data for a given game ID.
}
