% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{fetch_player_awards_data}
\alias{fetch_player_awards_data}
\title{Fetch Player Awards from API}
\usage{
fetch_player_awards_data(player_id)
}
\arguments{
\item{player_id}{A numeric or character vector representing the player's ID}
}
\value{
A list containing the raw player awards data.
}
\description{
Fetch Player Awards from API
}
