% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{fetch_player_career_data}
\alias{fetch_player_career_data}
\title{Fetch Player Career Data from API}
\usage{
fetch_player_career_data(player_id, per_mode)
}
\arguments{
\item{player_id}{A numeric or character vector representing the player's ID}

\item{per_mode}{A character string representing the aggregation method}
}
\value{
A list containing the raw player career data.
}
\description{
Fetch Player Career Data from API
}
