% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_dictionary.R
\name{fetch_player_dictionary_data}
\alias{fetch_player_dictionary_data}
\title{Fetch Player Dictionary from API}
\usage{
fetch_player_dictionary_data()
}
\value{
A list containing the raw player dictionary data.
}
\description{
Fetch Player Dictionary from API
}
