% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_news.R
\name{fetch_player_news}
\alias{fetch_player_news}
\title{Fetch Player News from Profile Page}
\usage{
fetch_player_news(url)
}
\arguments{
\item{url}{The NBA player's profile news page URL.}
}
\value{
A tibble containing the raw player news.
}
\description{
This function fetches player news for given player ID's.
}
