% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{fetch_player_profile_data}
\alias{fetch_player_profile_data}
\title{Fetch Player Profile Stats Data from API}
\usage{
fetch_player_profile_data(season_type, player_id, per_mode)
}
\arguments{
\item{season_type}{A character string specifying the type of season}

\item{player_id}{A numeric or character vector representing the player's ID}

\item{per_mode}{A character string representing the aggregation method}
}
\value{
A list containing the raw player profile stats data.
}
\description{
Fetch Player Profile Stats Data from API
}
