% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{fetch_player_splits_data}
\alias{fetch_player_splits_data}
\title{Fetch Player Splits Data from API}
\usage{
fetch_player_splits_data(
  season,
  season_type,
  measure_type,
  player_id,
  per_mode
)
}
\arguments{
\item{season}{A numeric value representing the season (e.g., 2024).}

\item{season_type}{A character string specifying the type of season}

\item{measure_type}{A character string specifying the measure type}

\item{player_id}{A numeric or character vector representing the player's ID}

\item{per_mode}{A character string representing the aggregation method}
}
\value{
A list containing the raw player career data.
}
\description{
Fetch Player Splits Data from API
}
