% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_stats.R
\name{fetch_player_stats}
\alias{fetch_player_stats}
\title{Fetch NBA Player Stats from API}
\usage{
fetch_player_stats(season, season_type, measure_type)
}
\arguments{
\item{season}{A numeric value representing the season (e.g., 2024).}

\item{season_type}{A character string specifying the type of season (e.g., "Regular Season").}

\item{measure_type}{A character string specifying the measure type (e.g., "Base").}
}
\value{
A data frame with cleaned NBA player stats.
}
\description{
This function fetches and cleans NBA player stats for a specified season and all measure types.
}
