% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{fetch_profile_info}
\alias{fetch_profile_info}
\title{Fetch Player Profile Information from API}
\usage{
fetch_profile_info(player_id)
}
\arguments{
\item{player_id}{A numeric or character vector representing the player's ID}
}
\value{
A data frame containing the raw player profile info.
}
\description{
This function fetches player profile info for given player ID's.
}
