% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_shots.R
\name{fetch_shots_data}
\alias{fetch_shots_data}
\title{Fetch NBA Shots from API}
\usage{
fetch_shots_data(season, season_type)
}
\arguments{
\item{season}{A numeric value representing the season (e.g., 2024).}

\item{season_type}{A character string specifying the type of season
(e.g., "Regular Season").}
}
\value{
A list containing raw shots data and league average data.
}
\description{
This function fetches raw NBA shot data and league average data for a given
season.
}
