% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_standings.R
\name{fetch_standings}
\alias{fetch_standings}
\title{Fetch and Process NBA Standings Data}
\usage{
fetch_standings(seasons)
}
\arguments{
\item{seasons}{A numeric value representing the year for which to fetch the standings data.}
}
\value{
A data frame containing the processed NBA standings data for the specified season.
}
\description{
This function fetches the NBA standings data for a specific season, processes it into a tidy
format, and returns it as a data frame.
}
