% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_dictionary.R
\name{fetch_team_dictionary_data}
\alias{fetch_team_dictionary_data}
\title{Fetch Team Dictionary from API}
\usage{
fetch_team_dictionary_data()
}
\value{
A list containing the raw team dictionary data.
}
\description{
Fetch Team Dictionary from API
}
