% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_win_probability.R
\name{fetch_win_probability_data}
\alias{fetch_win_probability_data}
\title{Fetch Win Probability Play-by-Play Data from API}
\usage{
fetch_win_probability_data(game_id)
}
\arguments{
\item{game_id}{The ID of the game for which to fetch data.}
}
\value{
A list containing the raw play-by-play data and metadata as tibbles.
}
\description{
This function fetches win probability play-by-play data for a given game ID.
}
