% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_headers_cdn}
\alias{generate_headers_cdn}
\title{Generate HTTP Headers for cdn.nba.com}
\usage{
generate_headers_cdn()
}
\value{
A named character vector of HTTP headers.
}
\description{
This function generates HTTP headers for the NBA Schedule API request.
}
