% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_headers_stats}
\alias{generate_headers_stats}
\title{Generate HTTP Headers for stats.nba.com}
\usage{
generate_headers_stats()
}
\value{
A named character vector of HTTP headers.
}
\description{
This function generates HTTP headers for stats API requests.
}
