% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_parameters_player_dict}
\alias{generate_parameters_player_dict}
\title{Generate Query Parameters for Player Dictionary API Requests}
\usage{
generate_parameters_player_dict()
}
\value{
A named list of query parameters.
}
\description{
This function generates the query parameters required for fetching the
NBA player dictionary
}
