% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_parameters_stats}
\alias{generate_parameters_stats}
\title{Generate Query Parameters for Stats API Requests}
\usage{
generate_parameters_stats(year, measure_type, season_type)
}
\arguments{
\item{year}{The season year for which to generate parameters.}

\item{measure_type}{The measure type for the API request.}

\item{season_type}{The season type for the API request.}
}
\value{
A named list of query parameters.
}
\description{
This function generates query parameters for the API request.
}
