% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_params_box_scores}
\alias{generate_params_box_scores}
\title{Generate Query Parameters for Box Score API Requests}
\usage{
generate_params_box_scores(game_id)
}
\arguments{
\item{game_id}{The game id for which to generate parameters.}
}
\value{
A named list of query parameters.
}
\description{
This function generates the query parameters required for fetching NBA box
scores data for a given season.
}
