% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_params_box_scores_fix}
\alias{generate_params_box_scores_fix}
\title{Generate Query Parameters for Box Score API Requests for PBP Fix}
\usage{
generate_params_box_scores_fix(game_id, start_range, end_range)
}
\arguments{
\item{game_id}{The game id for which to generate parameters.}

\item{start_range}{The start time for which to generate parameters.}

\item{end_range}{The end time for which to generate parameters.}
}
\value{
A named list of query parameters.
}
\description{
This function generates the query parameters required for fetching NBA box scores data for a
given season.
}
