% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_params_shots}
\alias{generate_params_shots}
\title{Generate Query Parameters for Shots API Requests}
\usage{
generate_params_shots(year, season_type)
}
\arguments{
\item{year}{The season year for which to generate parameters.}

\item{season_type}{The season type for the API request.}
}
\value{
A list of query parameters for the API request.
}
\description{
Generate Query Parameters for Shots API Requests
}
