% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_params_splits}
\alias{generate_params_splits}
\title{Generate Query Parameters for Player Splits API Requests}
\usage{
generate_params_splits(
  year,
  season_type = "Regular Season",
  measure_type = "Base",
  player_id,
  per_mode = "PerGame"
)
}
\arguments{
\item{year}{The season year for which to generate parameters.}

\item{season_type}{The season type for the API request.}

\item{measure_type}{The measure type for the API request.}

\item{player_id}{The player id for which to generate parameters.}

\item{per_mode}{A character string representing the aggregation method}
}
\value{
A named list of query parameters.
}
\description{
This function generates the query parameters required for fetching the
NBA player's splits page
}
