% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_params_standings}
\alias{generate_params_standings}
\title{Generate Query Parameters for Standings API Requests}
\usage{
generate_params_standings(year)
}
\arguments{
\item{year}{The season year for which to generate parameters.}
}
\value{
A named list of query parameters.
}
\description{
This function generates the query parameters required for fetching NBA
standings data for a given season.
}
