% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_params_team_dict}
\alias{generate_params_team_dict}
\title{Generate Query Parameters for Team Dictionary API Requests}
\usage{
generate_params_team_dict()
}
\value{
A named list of query parameters.
}
\description{
This function generates the query parameters required for fetching the
NBA team dictionary
}
