% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_all_players_for_league}
\alias{get_all_players_for_league}
\title{Fetch all players for a league from the pbpstats API}
\usage{
get_all_players_for_league(league = "nba")
}
\arguments{
\item{league}{Character. Which league to query; e.g. \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}
}
\value{
A data frame (\code{tibble}) of all players in the specified league.
}
\description{
A thin wrapper around the pbpstats API’s “get-all-players-for-league” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
