% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_assist_combo_summary}
\alias{get_assist_combo_summary}
\title{Fetch assist combo summary from the pbpstats API}
\usage{
get_assist_combo_summary(
  league = "nba",
  season,
  season_type = "Regular Season"
)
}
\arguments{
\item{league}{Character. Which league to query; e.g. \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g. \code{"2024-25"}.}

\item{season_type}{Character. One of \code{"Regular Season"}, \code{"Playoffs"}, or \code{"PlayIn"};
defaults to \code{"Regular Season"}.}
}
\value{
A tibble of assist combo summary as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-assist-combo-summary” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
