% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_assist_networks}
\alias{get_assist_networks}
\title{Fetch assist networks from the pbpstats API}
\usage{
get_assist_networks(
  league = "nba",
  season,
  season_type = "Regular Season",
  entity_type,
  entity_id,
  game_id = NULL
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g., \code{"2024-25"};
required for full-season calls; set to \code{NULL} if using \code{game_id}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}

\item{entity_type}{Character. One of:
\itemize{
\item \code{"Team"}
\item \code{"Lineup"}
}.}

\item{entity_id}{Integer or character. The pbpstats ID of the entity.}

\item{game_id}{Integer or character. The pbpstats ID of a specific game; if non-\code{NULL}, \code{season}
and \code{season_type} are ignored.}
}
\value{
A tibble of assist network data as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-assist-networks” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
