% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_current_season}
\alias{get_current_season}
\title{Get Current Season}
\usage{
get_current_season()
}
\value{
A character string representing the current season.
}
\description{
This function returns the current season in the format "YYYY-YY".
The season is determined based on the current date:
\itemize{
\item If the current date is before July 1st of the current year,
the season is "previous year - current year".
\item If the current date is after or on July 1st,
the season is "current year - next year".
}
}
