% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_game_stats}
\alias{get_game_stats}
\title{Fetch game stats from the pbpstats API}
\usage{
get_game_stats(game_id, type)
}
\arguments{
\item{game_id}{Integer or character. The pbpstats ID of the game.}

\item{type}{Character. One of:
\itemize{
\item \code{"Player"}
\item \code{"Lineup"}
\item \code{"LineupOpponent"}
}.}
}
\value{
A tibble of game statistics as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-game-stats” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
