% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_games}
\alias{get_games}
\title{Fetch games from the pbpstats API}
\usage{
get_games(season, season_type = "Regular Season")
}
\arguments{
\item{season}{Character or integer. The season to query; e.g. \code{"2024-25"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}
}
\value{
A tibble of games in the specified season as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-games” endpoint (NBA only).
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
