% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_live_game}
\alias{get_live_game}
\title{Fetch live game data from the pbpstats API}
\usage{
get_live_game(game_id, result_type)
}
\arguments{
\item{game_id}{Character or integer. The pbpstats ID of the game.}

\item{result_type}{Character. One of:
\itemize{
\item \code{"team"}
\item \code{"player"}
}.}
}
\value{
A tibble of live game data as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “live/game” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.  \url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
