% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_number_of_starter_stats_by_team}
\alias{get_number_of_starter_stats_by_team}
\title{Fetch number of starter stats by team from the pbpstats API}
\usage{
get_number_of_starter_stats_by_team(
  league = "nba",
  season,
  season_type = "Regular Season"
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g., \code{"2024-25"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}
}
\value{
A tibble of the number of starter stats by team as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-number-of-starter-stats-by-team” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
