% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_pace_efficiency_summary}
\alias{get_pace_efficiency_summary}
\title{Fetch pace efficiency summary from the pbpstats API}
\usage{
get_pace_efficiency_summary(
  league = "nba",
  season,
  season_type = "Regular Season",
  type
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g. \code{"2024-25"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}

\item{type}{Character. One of:
\itemize{
\item \code{"Team"}
\item \code{"Opponent"}
}.}
}
\value{
A tibble of pace efficiency summary as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-pace-efficiency-summary” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
