% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_player_headshot}
\alias{get_player_headshot}
\title{Generate NBA Player Headshot URL}
\usage{
get_player_headshot(player_id)
}
\arguments{
\item{player_id}{A numeric or character vector representing the player's ID.}
}
\value{
A character vector containing the headshot URLs.
}
\description{
Generates the URL for a player's headshot based on their NBA player ID.
}
