% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_player_measure_types}
\alias{get_player_measure_types}
\title{Get Available Measure Types for Player Stats}
\usage{
get_player_measure_types()
}
\value{
A character vector of measure types.
}
\description{
Returns a vector of available measure types for player stats.
}
