% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_playing_time_distribution}
\alias{get_playing_time_distribution}
\title{Fetch playing time distribution from the pbpstats API}
\usage{
get_playing_time_distribution(
  league = "nba",
  season,
  season_type = "Regular Season",
  team_id,
  player_id
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g., \code{"2024-25"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}

\item{team_id}{Integer or character. The pbpstats ID of the team.}

\item{player_id}{Integer or character. The pbpstats ID of the player.}
}
\value{
A tibble of playing time distribution as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-playing-time-distribution” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
