% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_possession_length_frequency}
\alias{get_possession_length_frequency}
\title{Fetch possession length frequency from the pbpstats API}
\usage{
get_possession_length_frequency(
  league = "nba",
  season,
  season_type = "Regular Season",
  poss_type,
  time_type,
  start_type = "All"
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g., \code{"2024-25"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}

\item{poss_type}{Character. One of:
\itemize{
\item \code{"Offense"}
\item \code{"Defense"}
}.}

\item{time_type}{Character. One of:
\itemize{
\item \code{"IncludeSecondChances"}
\item \code{"ExcludeSecondChances"}
}.}

\item{start_type}{Character. Defines the event that starts the possession; one of:
\itemize{
\item \code{"All"}
\item \code{"OffMissedFG"}
\item \code{"OffMissed2"}
\item \code{"OffMissed3"}
\item \code{"OffMadeFG"}
\item \code{"OffMade2"}
\item \code{"OffMade3"}
\item \code{"OffAtRimMake"}
\item \code{"OffAtRimMiss"}
\item \code{"OffAtRimBlock"}
\item \code{"OffShortMidRangeMake"}
\item \code{"OffShortMidRangeMiss"}
\item \code{"OffLongMidRangeMake"}
\item \code{"OffLongMidRangeMiss"}
\item \code{"OffArc3Make"}
\item \code{"OffArc3Miss"}
\item \code{"OffCorner3Make"}
\item \code{"OffCorner3Miss"}
\item \code{"OffFTMake"}
\item \code{"OffFTMiss"}
\item \code{"OffLiveBallTurnover"}
\item \code{"OffDeadball"}
\item \code{"OffTimeout"}
}.}
}
\value{
A tibble of possession length frequency as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-possession-length-frequency” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
