% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_score_margin_breakdown}
\alias{get_score_margin_breakdown}
\title{Fetch score margin breakdown from the pbpstats API}
\usage{
get_score_margin_breakdown(
  league = "nba",
  season,
  season_type = "Regular Season",
  team_id,
  period
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g., \code{"2024-25"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}

\item{team_id}{Integer or character. The pbpstats ID of the team.}

\item{period}{Character. One of:
\itemize{
\item \code{"All"}
\item \code{"1"}
\item \code{"2"}
\item \code{"3"}
\item \code{"4"}
\item \code{"SecondHalf"}
}.}
}
\value{
A tibble of score margin breakdown as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-score-margin-breakdown” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
