% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_play_by_play.R
\name{get_season_start_year}
\alias{get_season_start_year}
\title{Get Season Start Year}
\usage{
get_season_start_year(game_id)
}
\arguments{
\item{game_id}{The ID of the game for which to extract.}
}
\value{
A string of the season start year.
}
\description{
This function extracts the season start year from a game_id
}
