% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_shot_query_summary}
\alias{get_shot_query_summary}
\title{Fetch shot query summary from the pbpstats API}
\usage{
get_shot_query_summary(
  league = "nba",
  season,
  season_type = "Regular Season",
  type,
  from_margin,
  to_margin,
  from_time,
  to_time,
  period_gte = NULL,
  period_lte = NULL,
  period_equals = NULL
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g., \code{"2024-25"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}

\item{type}{Character. One of:
\itemize{
\item \code{"Team"}
\item \code{"Lineup"}
\item \code{"Opponent"}
\item \code{"LineupOpponent"}
}.}

\item{from_margin}{Numeric. Starting point margin for the breakdown.}

\item{to_margin}{Numeric. Ending point margin for the breakdown.}

\item{from_time}{Integer. Seconds remaining in the period at the start of a possession.}

\item{to_time}{Integer. Seconds remaining in the period at the end of a possession.}

\item{period_gte}{Integer. Filter for possessions in periods ≥ this value; defaults to \code{NULL}.}

\item{period_lte}{Integer. Filter for possessions in periods ≤ this value; defaults to \code{NULL}.}

\item{period_equals}{Integer. Filter for possessions in periods == this value;
defaults to \code{NULL}.}
}
\value{
A tibble of shot query summary data as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-shot-query-summary” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
