% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_shots}
\alias{get_shots}
\title{Fetch shots from the pbpstats API}
\usage{
get_shots(
  league = "nba",
  season,
  season_type = "Regular Season",
  entity_type,
  entity_id,
  start_type = "All",
  blocked,
  shot_value = NULL,
  oreb_shot_type = NULL,
  sec_since_oreb = NULL
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g. \code{"2024-25"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}

\item{entity_type}{Character. One of:
\itemize{
\item \code{"Player"}
\item \code{"Team"}
\item \code{"Lineup"}
\item \code{"LineupOpponent"}
}.}

\item{entity_id}{Integer or character. The pbpstats ID of the entity.}

\item{start_type}{Character. The event that started the possession; see endpoint docs for
options.}

\item{blocked}{Logical. Filter by blocked shots (\code{TRUE}) or unblocked (\code{FALSE}).}

\item{shot_value}{Integer. Shot value; \code{2} for two‐pointers or \code{3} for three‐pointers;
defaults to \code{NULL}.}

\item{oreb_shot_type}{Character. One of:
\itemize{
\item \code{"MissedFG"}
\item \code{"Missed2"}
\item \code{"MissedMidRange"}
\item \code{"Missed3"}
\item \code{"AtRim"}
\item \code{"AtRimBlocked"}
\item \code{"ShortMidRange"}
\item \code{"LongMidRange"}
\item \code{"Arc3"}
\item \code{"Corner3"}
\item \code{"FT"}
}; defaults to \code{NULL}.}

\item{sec_since_oreb}{Integer. Maximum seconds since an offensive
rebound (\code{SecondsSinceOrebLte}); defaults to \code{NULL}.}
}
\value{
A tibble of shot data as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-shots” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
