% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_team_logo}
\alias{get_team_logo}
\title{Generate NBA Team Logo URL}
\usage{
get_team_logo(team_id)
}
\arguments{
\item{team_id}{A numeric or character vector representing the team's ID.}
}
\value{
A character vector containing the logo URLs.
}
\description{
Generates the URL for a team's logo based on their NBA team ID.
}
