% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_team_measure_types}
\alias{get_team_measure_types}
\title{Get Available Measure Types for Team Stats}
\usage{
get_team_measure_types()
}
\value{
A character vector of measure types.
}
\description{
Returns a vector of available measure types for team stats.
}
