% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_teams}
\alias{get_teams}
\title{Fetch teams from the pbpstats API}
\usage{
get_teams(league = "nba")
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}
}
\value{
A tibble of teams in the specified league as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-teams” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
