% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_totals}
\alias{get_totals}
\title{Fetch totals from the pbpstats API}
\usage{
get_totals(
  league = "nba",
  season,
  season_type = "Regular Season",
  type,
  from_date = NULL,
  to_date = NULL,
  team_id = NULL,
  group_by = NULL,
  leverage = NULL,
  starter_state = NULL,
  start_type = NULL
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season}{Character or integer. The season to query; e.g. \code{"2024-25"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}

\item{type}{Character. One of:
\itemize{
\item \code{"Player"}
\item \code{"Team"}
\item \code{"Opponent"}
\item \code{"Lineup"}
\item \code{"LineupOpponent"}
}.}

\item{from_date}{Character or Date. Optional start date filter; defaults to \code{NULL}.}

\item{to_date}{Character or Date. Optional end date filter; defaults to \code{NULL}.}

\item{team_id}{Integer or character. Optional pbpstats team ID; defaults to \code{NULL}.}

\item{group_by}{Character. Optional grouping variable; e.g., \code{"Player"} or \code{"Team"};
defaults to \code{NULL}.}

\item{leverage}{Character. One or more leverage levels; one of:
\itemize{
\item \code{"VeryLow"}
\item \code{"Low"}
\item \code{"Medium"}
\item \code{"High"}
\item \code{"VeryHigh"}
}; comma‐separated for multiple values; defaults to \code{NULL}.}

\item{starter_state}{Character. One or more starter configurations; one of:
\itemize{
\item \code{"5v5"}
\item \code{"4v5"}
\item \code{"3v5"}
\item \code{"2v5"}
\item \code{"1v5"}
\item \code{"0v5"}
}; comma‐separated for multiple values; defaults to \code{NULL}.}

\item{start_type}{Character. One or more possession start types; one of:
\itemize{
\item \code{"All"}
\item \code{"OffMissedFG"}
\item \code{"OffMissed2"}
\item \code{"OffMissed3"}
\item \code{"OffMadeFG"}
\item \code{"OffMade2"}
\item \code{"OffMade3"}
\item \code{"OffAtRimMake"}
\item \code{"OffAtRimMiss"}
\item \code{"OffAtRimBlock"}
\item \code{"OffShortMidRangeMake"}
\item \code{"OffShortMidRangeMiss"}
\item \code{"OffLongMidRangeMake"}
\item \code{"OffLongMidRangeMiss"}
\item \code{"OffArc3Make"}
\item \code{"OffArc3Miss"}
\item \code{"OffCorner3Make"}
\item \code{"OffCorner3Miss"}
\item \code{"OffFTMake"}
\item \code{"OffFTMiss"}
\item \code{"OffLiveBallTurnover"}
\item \code{"OffDeadball"}
\item \code{"OffTimeout"}
}; comma‐separated for multiple values; defaults to \code{NULL}.}
}
\value{
A tibble of total metrics as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-totals” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
