% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbpstats.R
\name{get_year_over_year_plots}
\alias{get_year_over_year_plots}
\title{Fetch league year-over-year plots from the pbpstats API}
\usage{
get_year_over_year_plots(
  league = "nba",
  season_type = "Regular Season",
  left_axis,
  right_axis
)
}
\arguments{
\item{league}{Character. Which league to query; e.g., \code{"nba"} or \code{"wnba"}; defaults to \code{"nba"}.}

\item{season_type}{Character. One of:
\itemize{
\item \code{"Regular Season"}
\item \code{"Playoffs"}
\item \code{"PlayIn"}
\item \code{"All"}
}; defaults to \code{"Regular Season"}.}

\item{left_axis}{Character. One of the statistical metrics to plot on the left axis.
Options include:
\itemize{
\item \code{"PtsPer100Poss"}
\item \code{"SecondsPerPoss"}
\item \code{"FG3APct"}
\item \code{"Fg3Pct"}
\item \code{"AtRimFrequency"}
\item \code{"AtRimAccuracy"}
\item \code{"AtRimPctAssisted"}
\item \code{"ShortMidRangeFrequency"}
\item \code{"ShortMidRangeAccuracy"}
\item \code{"ShortMidRangePctAssisted"}
\item \code{"LongMidRangeFrequency"}
\item \code{"LongMidRangeAccuracy"}
\item \code{"LongMidRangePctAssisted"}
\item \code{"Corner3Frequency"}
\item \code{"Corner3Accuracy"}
\item \code{"Corner3PctAssisted"}
\item \code{"Arc3Frequency"}
\item \code{"Arc3Accuracy"}
\item \code{"Arc3PctAssisted"}
\item \code{"LiveBallTurnoverPct"}
\item \code{"EfgPct"}
\item \code{"DefFTReboundPct"}
\item \code{"OffFTReboundPct"}
\item \code{"DefTwoPtReboundPct"}
\item \code{"OffTwoPtReboundPct"}
\item \code{"DefThreePtReboundPct"}
\item \code{"OffThreePtReboundPct"}
\item \code{"DefFGReboundPct"}
\item \code{"OffFGReboundPct"}
\item \code{"OffAtRimReboundPct"}
\item \code{"OffShortMidRangeReboundPct"}
\item \code{"OffLongMidRangeReboundPct"}
\item \code{"OffArc3ReboundPct"}
\item \code{"OffCorner3ReboundPct"}
\item \code{"DefAtRimReboundPct"}
\item \code{"DefShortMidRangeReboundPct"}
\item \code{"DefLongMidRangeReboundPct"}
\item \code{"DefArc3ReboundPct"}
\item \code{"DefCorner3ReboundPct"}
\item \code{"SecondChancePtsPer100PossSecondChance"}
\item \code{"PenaltyPtsPer100PossPenalty"}
\item \code{"SecondChanceOffPossPer100Poss"}
\item \code{"FirstChancePtsPer100Poss"}
\item \code{"SecondChancePtsPer100Poss"}
\item \code{"PenaltyOffPossPer100Poss"}
\item \code{"Avg2ptShotDistance"}
\item \code{"Avg3ptShotDistance"}
}}

\item{right_axis}{Character. One of the same options as \code{left_axis}, plotted on the right axis.}
}
\value{
A tibble of year-over-year league metrics suitable for plotting, as returned by the API.
}
\description{
A thin wrapper around the pbpstats API’s “get-league-year-over-year-plots” endpoint.
}
\references{
Darryl Blackport (2020). PBPStats API, MIT License.
\url{https://api.pbpstats.com/docs}
}
\seealso{
\url{https://api.pbpstats.com/docs}
}
