% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp.R
\name{identify_fg_and_one}
\alias{identify_fg_and_one}
\title{Identify Field Goals of an And-One}
\usage{
identify_fg_and_one(data)
}
\arguments{
\item{data}{A data frame with raw play-by-play data}
}
\value{
A data frame with the only the fgs of an and-one.
}
\description{
This function identifies field goals made as part of an "and-one".
}
