% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp.R
\name{initial_playtime}
\alias{initial_playtime}
\title{Calculate Initial Lineup Playtime}
\usage{
initial_playtime(data, location)
}
\arguments{
\item{data}{A data frame containing game event data, including lineup
information, periods, and time.}

\item{location}{A string indicating the location ("home" or "away") to
specify which team's lineup to calculate.}
}
\value{
A tibble with the initial dataframe containing playtime data.
}
\description{
This function creates the initial playtime dataframe that calculates the
playtime for each lineup.
}
