% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_data_frames}
\alias{join_data_frames}
\title{Join Multiple Data Frames by Removing Duplicate Columns}
\usage{
join_data_frames(data_frames, join_columns)
}
\arguments{
\item{data_frames}{A list of data frames to be joined.}

\item{join_columns}{A character vector of column names to join the data
frames by.}
}
\value{
A single data frame with all the data frames joined by the specified
columns.
}
\description{
This function takes a list of data frames and joins them using \code{full_join},
ensuring that only unique columns from each data frame are included in the
final result. It removes any columns that already exist in the preceding data
frames.
}
