% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{join_scores}
\alias{join_scores}
\title{Join Scores}
\usage{
join_scores(all_stats, team_games, opp_team_games)
}
\arguments{
\item{all_stats}{A data frame with combined team and opponent stats.}

\item{team_games}{A data frame with processed team game data.}

\item{opp_team_games}{A data frame with processed opponent team game data.}
}
\value{
A data frame with joined stats.
}
\description{
This function joins team and opponent stats with game information.
}
