% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp.R
\name{make_pbp_corrections}
\alias{make_pbp_corrections}
\title{Play-by-Play Corrections}
\usage{
make_pbp_corrections(data)
}
\arguments{
\item{data}{The raw play-by-play data.}
}
\value{
A data frame with corrections made to events and descriptions
}
\description{
This function handles data corrections in play-by-play data.
}
