% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{nba_career_rankings}
\alias{nba_career_rankings}
\title{Get Player Career Rankings}
\usage{
nba_career_rankings(player_id, per_mode = "PerGame")
}
\arguments{
\item{player_id}{A numeric or character vector representing the player's ID}

\item{per_mode}{A character string representing the aggregation method}
}
\value{
A tibble with all season rankings
}
\description{
This function gets the player's season rankings for regular and post-season.
}
