% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{nba_career_splits}
\alias{nba_career_splits}
\title{Get Player Splits}
\usage{
nba_career_splits(
  season,
  season_type = "Regular Season",
  measure_type = "Base",
  player_id,
  per_mode = "PerGame"
)
}
\arguments{
\item{season}{A numeric value representing the season (e.g., 2024).}

\item{season_type}{A character string specifying the type of season
(default = "Regular Season"). Valid options include:
\itemize{
\item \strong{"Pre Season"} - Pre Season games.
\item \strong{"Regular Season"} - Regular Season games.
\item \strong{"Playoffs"} - Playoff games.
\item \strong{"All Star"} - All Star games.
\item \strong{"IST"} - NBA Cup games.
\item \strong{"PlayIn"} - Play In games.
}}

\item{measure_type}{A character vector specifying the types of stats
(default = "Base"). Valid options include:
\itemize{
\item \strong{"Base"} - Traditional stats.
\item \strong{"Advanced"} - Advanced stats.
\item \strong{"Usage"} - Usage stats.
\item \strong{"Misc"} - Misc stats.
\item \strong{"Scoring"} - Scoring stats.
}}

\item{player_id}{A numeric or character vector representing the player's ID}

\item{per_mode}{A character string representing the aggregation method
(default = "PerGame") Valid options include:
\itemize{
\item \strong{"PerGame"} - Stats on per game basis.
\item \strong{"Totals"} - Stat totals.
}}
}
\value{
A tibble with all splits totals
}
\description{
This function gets the player's various splits by measure type and season.
}
