% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{nba_career_stats}
\alias{nba_career_stats}
\title{Get Player Career Totals}
\usage{
nba_career_stats(player_id, per_mode = "PerGame")
}
\arguments{
\item{player_id}{A numeric or character vector representing the player's ID}

\item{per_mode}{A character string representing the aggregation method}
}
\value{
A tibble with all career totals
}
\description{
This function gets the player's regular season, post-season, all-star, and preseason career
totals.
}
