% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{nba_player_awards}
\alias{nba_player_awards}
\title{Get NBA Player Awards History}
\usage{
nba_player_awards(player_id)
}
\arguments{
\item{player_id}{A numeric or character vector representing the player's ID.}
}
\value{
A character vector containing the headshot URLs.
}
\description{
Gets the award history for a player based on their NBA player ID.
}
