% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_dictionary.R
\name{nba_player_dictionary}
\alias{nba_player_dictionary}
\title{Get NBA Player Dictionary}
\usage{
nba_player_dictionary()
}
\value{
A tibble containing the player dictionary.
}
\description{
This function gets NBA player dictionary data and returns a tibble..
}
