% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_dictionary.R
\name{nba_player_headshot}
\alias{nba_player_headshot}
\title{Get NBA Player Headshot Photo}
\usage{
nba_player_headshot(player_id)
}
\arguments{
\item{player_id}{A numeric or character vector representing the player's ID.}
}
\value{
A character vector containing the headshot URLs.
}
\description{
Gets the URL for a player's headshot based on their NBA player ID.
}
