% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{nba_player_highs}
\alias{nba_player_highs}
\title{Get Player Highs}
\usage{
nba_player_highs(player_id)
}
\arguments{
\item{player_id}{A numeric or character vector representing the player's ID}
}
\value{
A tibble with season and career highs.
}
\description{
This function gets the player's current season and career highs.
}
