% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nba_player_lookup}
\alias{nba_player_lookup}
\title{NBA Player Lookup}
\usage{
nba_player_lookup()
}
\value{
A tibble with person_id and player_name.
}
\description{
This function retrieves a simplified player dictionary from the NBA API, including player IDs
and full names.
}
