% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_news.R
\name{nba_player_news}
\alias{nba_player_news}
\title{Get NBA Player News}
\usage{
nba_player_news(player_id)
}
\arguments{
\item{player_id}{A single or vector of player IDs to fetch player news for.}
}
\value{
A tibble containing the processed player news for the player(s).
}
\description{
Retrieves and processes player news for a single or multiple players.
}
