% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_stats.R
\name{nba_player_stats}
\alias{nba_player_stats}
\title{Get NBA Player Stats}
\usage{
nba_player_stats(seasons, season_type = "Regular Season")
}
\arguments{
\item{seasons}{A numeric vector of seasons (e.g., 2024) for which to scrape NBA player stats.}

\item{season_type}{The season type for the API request.}
}
\value{
A named list where each element is a data frame containing player level stats for that season.
}
\description{
This function gets NBA player stats for the specified seasons returning a
list where each season contains a data frame for each measure type. Function
pauses for five seconds after each season to prevent timeout issues.
}
