% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{nba_profile_info}
\alias{nba_profile_info}
\title{Get NBA Player Profile Descriptive Information}
\usage{
nba_profile_info(player_id)
}
\arguments{
\item{player_id}{A single or vector of player IDs to fetch profile information for.}
}
\value{
A tibble containing the processed profile information for the player(s).
}
\description{
Retrieves and processes profile information for a single or multiple players
}
