% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_profiles.R
\name{nba_profile_stats}
\alias{nba_profile_stats}
\title{Get Player Profile Stats}
\usage{
nba_profile_stats(
  season_type = "Regular Season",
  player_id,
  per_mode = "PerGame"
)
}
\arguments{
\item{season_type}{A character string specifying the type of season
(default = "Regular Season"). Valid options include:
\itemize{
\item \strong{"Pre Season"} - Pre Season games.
\item \strong{"Regular Season"} - Regular Season games.
\item \strong{"Playoffs"} - Playoff games.
\item \strong{"All Star"} - All Star games.
\item \strong{"IST"} - NBA Cup games.
\item \strong{"PlayIn"} - Play In games.
}}

\item{player_id}{A numeric or character vector representing the player's ID}

\item{per_mode}{A character string representing the aggregation method
(default = "PerGame") Valid options include:
\itemize{
\item \strong{"PerGame"} - Stats on per game basis.
\item \strong{"Totals"} - Stat totals.
}}
}
\value{
A list of tibbles with all profile stats
}
\description{
This function gets the player's various stats by measure type and
aggregation method.
}
