% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_schedule.R
\name{nba_schedule}
\alias{nba_schedule}
\title{Get NBA Schedule Data}
\usage{
nba_schedule()
}
\value{
A data frame with processed NBA schedule.
}
\description{
This function gets the current NBA schedule, processes it, and returns it.
}
