% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{nba_scores}
\alias{nba_scores}
\title{Get NBA Scores}
\usage{
nba_scores(seasons, season_type = "Regular Season")
}
\arguments{
\item{seasons}{A numeric vector of seasons (e.g., 2024) for which to scrape
NBA scores.}

\item{season_type}{A character string specifying the type of season
(e.g., "Regular Season").}
}
\value{
A data frame containing the NBA scores and related statistics.
}
\description{
This function gets NBA scores for the specified seasons. Function pauses for
five seconds after each season to prevent timeout issues.
}
