% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_shots.R
\name{nba_shots}
\alias{nba_shots}
\title{Get NBA Shots}
\usage{
nba_shots(seasons, season_type = "Regular Season", return_nested = FALSE)
}
\arguments{
\item{seasons}{A numeric vector of seasons (e.g., 2024) for which to fetch
NBA shot data.}

\item{season_type}{A character string specifying the type of season. Valid
options include:
\itemize{
\item \strong{"Pre Season"} - Pre Season games.
\item \strong{"Regular Season"} - Regular Season games.
\item \strong{"Playoffs"} - Playoff games.
\item \strong{"All Star"} - All Star games.
\item \strong{"IST"} - NBA Cup games.
\item \strong{"PlayIn"} - Play In games.
}}

\item{return_nested}{A logical value. If FALSE (default), returns a single
combined data frame for all seasons.If TRUE, returns a list of data frames,
one for each season.}
}
\value{
A named a data frame containing NBA shots data for the specified
seasons.
}
\description{
This function gets NBA shot data for the specified seasons and
returns a data frame. Function pauses for five seconds after each season to
prevent timeout issues.
}
