% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_standings.R
\name{nba_standings}
\alias{nba_standings}
\title{Get NBA Standings Data}
\usage{
nba_standings(seasons, return_nested = FALSE)
}
\arguments{
\item{seasons}{A numeric vector of seasons for which to get the standings data.}

\item{return_nested}{A logical value. If FALSE (default), returns a single
combined data frame for all seasons. If TRUE, returns a list of data frames, one for each season.}
}
\value{
A data frame containing the NBA standings data for the specified seasons.
}
\description{
This function gets NBA standings as a data frame for a single or multiple seasons.
}
