% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_dictionary.R
\name{nba_team_dictionary}
\alias{nba_team_dictionary}
\title{Get NBA Team Dictionary}
\usage{
nba_team_dictionary()
}
\value{
A tibble containing the team dictionary.
}
\description{
This function gets NBA team dictionary data and returns a tibble.
}
