% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_dictionary.R
\name{nba_team_logo}
\alias{nba_team_logo}
\title{Get NBA Team Logo}
\usage{
nba_team_logo(team_id)
}
\arguments{
\item{team_id}{A numeric or character vector representing the team's ID.}
}
\value{
A character vector containing the logo URLs.
}
\description{
Gets the URL for a team's logo based on their NBA team ID.
}
