% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_team_stats.R
\name{nba_team_stats}
\alias{nba_team_stats}
\title{Get NBA Team Stats}
\usage{
nba_team_stats(seasons, season_type = "Regular Season")
}
\arguments{
\item{seasons}{A numeric vector of seasons (e.g., 2024) for which to scrape
NBA team stats.}

\item{season_type}{The season type for the API request.}
}
\value{
A named list where each element is a data frame containing team level
stats for that season.
}
\description{
This function gets NBA team stats for the specified seasons returning a list
where each season contains a data frame for each measure type. Function
pauses for five seconds after each season to prevent timeout issues.
}
