% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp.R
\name{pbp_box_scores}
\alias{pbp_box_scores}
\title{Calculate Play-by-Play Box Scores for Multiple Games}
\usage{
pbp_box_scores(game_df)
}
\arguments{
\item{game_df}{A data frame of play-by-play data processed with
\code{pbp_lineups()}}
}
\value{
A tibble containing box scores for all lineups across multiple games
}
\description{
This function processes play-by-play box scores for multiple games.
IMPORTANT: This function can only be used after \code{pbp_lineups()} has
been applied to the play-by-play data to add lineup columns.
}
