% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_play_by_play.R
\docType{data}
\name{pbp_column_map}
\alias{pbp_column_map}
\title{PBP Column Name Map}
\format{
An object of class \code{character} of length 23.
}
\usage{
pbp_column_map
}
\description{
This vector is used to rename data.nba.com PBP columns.
}
\keyword{datasets}
