% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_types.R
\docType{data}
\name{player_measure_types}
\alias{player_measure_types}
\title{Measure Types for NBA Player Stats}
\format{
A character vector.
}
\source{
FastBreakR
}
\usage{
player_measure_types
}
\description{
This file contains the available measure types for NBA player statistics.
The measure types can be used to filter or categorize player stats in various analyses and
models.
}
\keyword{datasets}
