% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_processing.R
\name{players_in_events}
\alias{players_in_events}
\title{Track Players Involved in Events But Not Substitutions}
\usage{
players_in_events(data, subs_made)
}
\arguments{
\item{data}{A data frame containing game event data.}

\item{subs_made}{A data frame containing subs made}
}
\value{
A data frame with players involved in events but not subbed
}
\description{
This function processes the input data to identify players involved in events
(excluding substitutions) and ensures that only players who are not substituted are included in
the final output.
}
