% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_schedule.R
\name{process_nba_schedule}
\alias{process_nba_schedule}
\title{Process NBA Schedule Data}
\usage{
process_nba_schedule(data)
}
\arguments{
\item{data}{A list containing the fetched NBA schedule data.}
}
\value{
A data frame with processed NBA schedule.
}
\description{
This function processes the fetched NBA schedule data by adding game and schedule details.
}
