% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{process_opp_all_stats}
\alias{process_opp_all_stats}
\title{Process Opponent All Stats}
\usage{
process_opp_all_stats(team_all_stats)
}
\arguments{
\item{team_all_stats}{A data frame containing processed team stats.}
}
\value{
A data frame with processed opponent team stats.
}
\description{
This function processes opponent team stats data.
}
