% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{process_opp_team_games}
\alias{process_opp_team_games}
\title{Process Opponent Team Games}
\usage{
process_opp_team_games(team_games)
}
\arguments{
\item{team_games}{A data frame containing processed team game data.}
}
\value{
A data frame with processed opponent team game data.
}
\description{
This function processes opponent team game data.
}
