% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_player_stats.R
\name{process_player_measures}
\alias{process_player_measures}
\title{Process Player Data}
\usage{
process_player_measures(player_data)
}
\arguments{
\item{player_data}{A data frame containing player stats to be converted to numeric.}
}
\value{
A data frame with the \code{game_date} column converted to Date and other relevant columns
converted to numeric.
}
\description{
This helper function processes player data by converting the \code{game_date} column to a date
object and converting necessary columns to numeric values.
}
