% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_shots.R
\name{process_shots}
\alias{process_shots}
\title{Process NBA Shots Data}
\usage{
process_shots(data)
}
\arguments{
\item{data}{A named list or data frame with shots or league average data.}
}
\value{
A named list or data frame, whichever was provided as input.
}
\description{
To be used on the output of \code{nba_shots()} or \code{nba_shots_league_avg()} within
the \code{nba_shot_chart()} function.
Processes shot and league average data within each NBA season.
Applies transformations to prepare the data for analysis and visualization.
}
