% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{process_team_games}
\alias{process_team_games}
\title{Process Team Games}
\usage{
process_team_games(team_all_stats)
}
\arguments{
\item{team_all_stats}{A data frame containing processed team stats.}
}
\value{
A data frame with processed team game data.
}
\description{
This function processes team game data.
}
