% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_team_stats.R
\name{process_team_measures}
\alias{process_team_measures}
\title{Process Team Data}
\usage{
process_team_measures(team_data)
}
\arguments{
\item{team_data}{A data frame containing team stats to be converted to
numeric.}
}
\value{
A data frame with the \code{game_date} column converted to Date and other
relevant columns converted to numeric.
}
\description{
This helper function processes team data by converting the \code{game_date} column
to a Date object and converting necessary columns to numeric values.
}
