% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_win_probability.R
\name{process_win_probability_data}
\alias{process_win_probability_data}
\title{Process Win Probability Play-by-Play Data}
\usage{
process_win_probability_data(data, metadata)
}
\arguments{
\item{data}{Raw win probability play-by-play data.}

\item{metadata}{Metadata associated with the data.}
}
\value{
A tibble with processed win probability play-by-play data.
}
\description{
This function processes the raw win probability play-by-play data.
}
