% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seconds_passed}
\alias{seconds_passed}
\title{Calculate Seconds Passed in Game}
\usage{
seconds_passed(time_str, period)
}
\arguments{
\item{time_str}{A string representing the time in the current period.}

\item{period}{An integer representing the period.}
}
\value{
A numeric value representing the total seconds passed in the game.
}
\description{
Calculates the total seconds passed in the game based on time and period.
}
