% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_processing.R
\name{subs_in_quarter}
\alias{subs_in_quarter}
\title{Track Player Substitutions in Each Quarter}
\usage{
subs_in_quarter(data)
}
\arguments{
\item{data}{A data frame containing game event data with substitution events.}
}
\value{
A data frame with the players subbed during a quarter
}
\description{
This function processes the input data to identify player substitutions
to show who was substituted in and out during each quarter.
}
