% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_types.R
\docType{data}
\name{team_measure_types}
\alias{team_measure_types}
\title{Measure Types for NBA Team Stats}
\format{
A character vector.
}
\source{
FastBreakR
}
\usage{
team_measure_types
}
\description{
This file contains the available measure types for NBA team statistics.
The measure types can be used to filter or categorize team stats in various
analyses and models.
}
\keyword{datasets}
