% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_processing.R
\name{track_pbp_lineups}
\alias{track_pbp_lineups}
\title{Process Play-by-Play Data to Track Lineups for Each Play}
\usage{
track_pbp_lineups(data, quarter_start_lineups, all_subs, fgs_data)
}
\arguments{
\item{data}{A data frame with raw play-by-play data}

\item{quarter_start_lineups}{A data frame with lineups to begin each quarter}

\item{all_subs}{A data frame with all subs made}

\item{fgs_data}{A data frame with fgs before and-one}
}
\value{
A data frame with the lineups for all play-by-play data
}
\description{
This function processes play-by-play data to track lineups. It adds details
about the starting lineups, substitutions, and final lineups for each
game period. The function also identifies free throw sequences and
tracks the players involved in those events.
}
\details{
The resulting data frame includes four additional columns:
\describe{
\item{\code{lineup_away_pt}}{Tracks the total playtime for away lineups.}
\item{\code{lineup_home_pt}}{Tracks the total playtime for home lineups.}
\item{\code{lineup_away}}{Tracks the away lineup for calculating stats.}
\item{\code{lineup_home}}{Tracks the home lineup for calculating stats.}
}
}
