% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_processing.R
\name{track_subs_and_lineups}
\alias{track_subs_and_lineups}
\title{Track Player Substitutions and Lineups}
\usage{
track_subs_and_lineups(data, quarter_start_lineups)
}
\arguments{
\item{data}{A data frame with raw play-by-play data}

\item{quarter_start_lineups}{A data frame with lineups to begin each quarter}
}
\value{
A data frame with all subs made
}
\description{
This function processes play-by-play data to track player substitutions
and the lineups before and after each substitution. It provides a breakdown
of player substitutions by game and period, along with the updated lineups
for both home and away teams.
}
