% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pbp.R
\name{calculate_lineup_stats}
\alias{calculate_lineup_stats}
\title{Calculate Lineup Base Stats}
\usage{
calculate_lineup_stats(game_df, lineup_playtime)
}
\arguments{
\item{game_df}{A data frame containing play-by-play data with lineups.}

\item{lineup_playtime}{A data frame containing lineups playtime.}
}
\value{
A tibble containing base stats data.
}
\description{
This function calculates the base stats for each lineup to be used in
calculating play-by-play box scores.
}
