% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{nba_scores}
\alias{nba_scores}
\title{Get NBA Scores}
\usage{
nba_scores(seasons, season_type = "Regular Season", return_nested = FALSE)
}
\arguments{
\item{seasons}{A numeric vector of seasons (e.g., 2024) for which to fetch NBA scores.}

\item{season_type}{A character string specifying the type of season.
Valid options include:
\itemize{
\item \strong{"Pre Season"} - Pre Season games.
\item \strong{"Regular Season"} - Regular Season games.
\item \strong{"Playoffs"} - Playoff games.
\item \strong{"All Star"} - All Star games.
\item \strong{"IST"} - NBA Cup games.
\item \strong{"PlayIn"} - Play In games.
}}

\item{return_nested}{A logical value. If FALSE (default), returns a single
combined data frame for all seasons. If TRUE, returns a list of data frames, one for each season.}
}
\value{
A data frame containing the NBA scores data for the specified seasons.
}
\description{
This function gets NBA scores for the specified seasons. Function pauses for
five seconds after each season to prevent timeout issues.
}
