% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_shots.R
\name{nba_shots}
\alias{nba_shots}
\title{Get NBA Shots}
\usage{
nba_shots(seasons, season_type = "Regular Season")
}
\arguments{
\item{seasons}{A numeric vector of seasons (e.g., 2024) for which to scrape
NBA shot data.}

\item{season_type}{A character string specifying the type of season
(e.g., "Regular Season").}
}
\value{
A named list where each element is a list containing two data frames:
\code{shots} and \code{league_avg}.
}
\description{
This function gets NBA shot data for the specified seasons and
returns a list where each season contains two data frames: \code{shots} and
\code{league_avg}. Function pauses for five seconds after each season to prevent
timeout issues.
}
