% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_standings.R
\name{nba_standings}
\alias{nba_standings}
\title{Get NBA Standings Data}
\usage{
nba_standings(seasons)
}
\arguments{
\item{seasons}{A numeric vector of seasons for which to get the standings data.}
}
\value{
A named list of data frames, each containing the NBA standings data
for the specified seasons.
}
\description{
This function gets NBA standings data for a single or multiple seasons
and returns it as a list of data frames.
}
