% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_shots.R
\name{process_shots}
\alias{process_shots}
\title{Process NBA Shots Data (function under development)}
\usage{
process_shots(data)
}
\arguments{
\item{data}{A named list where each element represents a season. Each season
contains a list with two data frames: \code{shots} and \code{league_avg}.}
}
\value{
A named list of processed data frames for each season. Transforms the
data to prepare for analysis.
}
\description{
This function is currently in development and not exported.
}
\details{
To be used on the output of \code{nba_shots()}. Processes shot and league average
data frames within each NBA season. Applies transformations to prepare the
data for analysis.
}
